+++ src/mesher/geometry/abc.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import abc
-import pathlib
 from typing import Iterable, Literal
 
 
@@ -58,20 +56,6 @@ class IPoint(metaclass=abc.ABCMeta):
     @abc.abstractmethod
     def y(self) -> float:
         """This gets the y-position of the point."""
-
-        ...
-
-
-class IReader(metaclass=abc.ABCMeta):
-
-    @staticmethod
-    @abc.abstractmethod
-    def read(filename: str | pathlib.Path) -> list[IRing]:
-        """
-        Read point data from file and load it into figures.
-
-        TODO: update this when polygons/cells are implemented.
-        """
 
         ...
 
@@ -203,18 +187,4 @@ class IRing(metaclass=abc.ABCMeta):
         """This splits a ring that has self-intersections into multiple,
         non-self-intersecting rings."""
 
-        ...
-
-
-class IWriter(metaclass=abc.ABCMeta):
-
-    @staticmethod
-    @abc.abstractmethod
-    def write(filename: str | pathlib.Path, data: dict[str, IRing]) -> None:
-        """
-        Write point data from figures to file.
-
-        TODO: update this when polygons/cells are implemented.
-        """
-        
         ...
